# server-application/restapi/src/main/resources/openapi-specification.yml
openapi: 3.0.3
info:
  title: Mobile NotWork Game REST API
  version: "1.0"
...
paths:
  /spieler:
    get:
      summary: Gibt einen Spieler zu gegebener sid zurueck oder listet alle Spieler auf.
      description: Gibt einen Spieler zu gegebener sid zurueck oder listet alle Spieler auf.
      tags:
        - SpielerRESTController
      security:
        - spielerTokenAuth: []
      operationId: getSpieler
      parameters:
        - name: sid
          in: query
          description: Die Spieler ID
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Erfolgreiche Ermittlung der Spieler
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spieler'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /spieler/register:
    post:
      summary: Spieler registrieren
      description: Registrierung eines Spielers gemaess LF02.
      tags:
        - SpielerRESTController
      operationId: spielerRegistrieren
      parameters:
        - name: geraeteID
          in: query
          description: Die ID des Geraets, mit dem sich der Spieler registriert.
          required: true
          schema:
            type: string
        - name: einwilligungAGB
          in: query
          description: Zeigt an, ob der Spieler die AGBs akzeptiert hat.
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Erfolgreiche Registrierung
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Fehlgeschlagene Registrierung
          content:
            application/json:
              schema:
                type: string
  /spieler/login:
    post:
      summary: Spieler authentifizieren und autorisieren
      description: Authentifizierung und Autorisierung eines Spielers gemaess LF01.
      operationId: spielerLogin
      tags:
        - SpielerRESTController
      parameters:
        - name: geraeteID
          in: query
          description: Die ID des Geraets, mit dem sich der Spieler registriert.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Erfolgreiche Authentifizierung und Autorisierung
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginReplyDaten'
        '401':
          description: Ungueltige Anmeldeinformationen
          content:
            application/json:
              schema:
                type: string
        '403':
          description: Spieler blockiert
          content:
            application/json:
              schema:
                type: string
  /spieler/benutzername/{sid}:
    get:
      summary: Benutzernamen fuer Spieler ermitteln
      description: Fuer eine gegebene sid den zugehoerigen Benutzernamen ermitteln
      operationId: getBenutzername
      tags:
        - SpielerRESTController
      parameters:
        - name: sid
          in: path
          description: Die ID des Spielers.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Benutzername erfolgreich ermittelt
          content:
            application/json:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /spieler/benutzername-aendern/{sid}:
    post:
      summary: Benutzernamen aendern
      description: Aenderung des Benutzernamens gemaess LF06.
      operationId: spielerBenutzernameAendern
      tags:
        - SpielerRESTController
      security:
        - spielerTokenAuth: []
      parameters:
        - name: sid
          in: path
          description: Die ID des Spielers, dessen Benutzername geaendert werden soll.
          required: true
          schema:
            type: integer
        - name: neuerBenutzername
          in: query
          description: Der neue Benutzername.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Erfolgreiche Aenderung des Benutzernamens
          content:
            application/json:
              schema:
                type: string
                example: "spieler123"
        '400':
          description: Fehlgeschlagene Aenderung des Benutzernamens
          content:
            application/json:
              schema:
                type: string
                example: "Ungueltige Anfrage, fehlende oder ungueltige sid oder neuerBenutzername"
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                type: string
                example: "Authentifizierung fehlgeschlagen, bitte melden Sie sich erneut an."
        '403':
          description: Benutzername bereits vergeben
          content:
            application/json:
              schema:
                type: string
                example: "Der gewuenschte Benutzername ist bereits vergeben, bitte waehlen Sie einen anderen."
  /spieler/blockieren/{sid}:
    post:
      summary: Spieler blockieren
      description: Blockieren eines Spielers gemaess LF19.
      operationId: spielerBlockieren
      tags:
        - SpielerRESTController
      security:
        - adminTokenAuth: []

      parameters:
        - name: sid
          in: path
          description: Die ID des Spielers, der blockiert werden soll.
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: Spieler erfolgreich blockiert
        '400':
          description: Fehlgeschlagene Blockierung des Spielers
          content:
            application/json:
              schema:
                type: string
        '401':
          $ref: "#/components/responses/NotAuthenticated"
  /spieler/freischalten/{sid}:
    post:
      summary: Spieler freischalten
      description: Freischalten eines Spielers gemaess LF20.
      operationId: spielerFreischalten
      tags:
        - SpielerRESTController
      security:
        - adminTokenAuth: []

      parameters:
        - name: sid
          in: path
          description: Die ID des Spielers, der freigeschaltet werden soll.
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: Spieler erfolgreich freigeschaltet
        '400':
          description: Fehlgeschlagene Freischaltung des Spielers
          content:
            application/json:
              schema:
                type: string
        '401':
          $ref: "#/components/responses/NotAuthenticated"
...
  /messung:
    post:
      summary: Messung erfassen
      description: Erfassen einer Messung gemaess LF08.
      operationId: messungErfassen
      tags:
        - MessungRESTController
      security:
        - spielerTokenAuth: []

      parameters:
        - name: sid
          in: query
          description: Die ID des Spielers, fuer den die Messung erfasst wird.
          required: true
          schema:
            type: integer
        - name: ksid
          in: query
          description: Die Kartensegment ID.
          required: true
          schema:
            type: integer
        - name: mfid
          in: query
          description: Die Mobilfunkanbieter ID.
          required: true
          schema:
            type: integer
        - name: msid
          in: query
          description: Die Mobilfukstandard ID.
          required: true
          schema:
            type: integer
        - name: rssi
          in: query
          description: Der Received Signal Strength Indicator (RSSI).
          required: true
          schema:
            type: integer
        - name: rtt
          in: query
          description: Die Round-Trip Time (RTT).
          required: true
          schema:
            type: integer
        - name: snr
          in: query
          description: Das Signal-to-Noise Ratio (SNR).
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messung erfolgreich erfasst
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessungReply'
        '400':
          description: Fehlgeschlagene Erfassung der Messung
        '404':
          description: Messung konnte keinem Kartensegment zugeordnet werden
        '401':
          $ref: '#/components/responses/NotAuthenticated'
...
components:
  schemas:
    LoginReplyDaten:
      type: object
      properties:
        id:
          type: integer
          format: int32
        token:
          type: string
    Spieler:
      type: object
      properties:
        sid:
          type: integer
          format: int32
        benutzername:
          type: string
        geraeteid:
          type: string
        einwilligungagb:
          type: boolean
...
    MessungReply:
      type: object
      properties:
        punkte:
          type: integer
          format: int32
...
  securitySchemes:
    spielerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            type: string
            example: "Angeforderte Ressource wurde nicht gefunden."
    NotAuthenticated:
      description: Nicht authentifiziert
      content:
        application/json:
          schema:
            type: string
            example: "Nicht authentifiziert, bitte melden Sie sich an."